LayoutTests/fast/events/constructors/keyboard-event-constructor.html

<!DOCTYPE html>
<html>
<head>
<script src="../../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>

description("This tests the constructor for the KeyboardEvent DOM class.");

var testObject = {nyannyan: 123};

// No initializer is passed.
shouldBe("new KeyboardEvent('eventType').bubbles", "false");
shouldBe("new KeyboardEvent('eventType').cancelable", "false");
shouldBe("new KeyboardEvent('eventType').view", "null");
shouldBe("new KeyboardEvent('eventType').detail", "0");

// bubbles is passed.
shouldBe("new KeyboardEvent('eventType', { bubbles: false }).bubbles", "false");
shouldBe("new KeyboardEvent('eventType', { bubbles: true }).bubbles", "true");

// cancelable is passed.
shouldBe("new KeyboardEvent('eventType', { cancelable: false }).cancelable", "false");
shouldBe("new KeyboardEvent('eventType', { cancelable: true }).cancelable", "true");

// view is passed.
// Window objects.
shouldBe("new KeyboardEvent('eventType', { view: window }).view", "window");
shouldBe("new KeyboardEvent('eventType', { view: this }).view", "this");

// Non-window objects.
shouldBe("new KeyboardEvent('eventType', { view: testObject }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: document }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: undefined }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: null }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: false }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: true }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: '' }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: 'chocolate' }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: 12345 }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: 18446744073709551615 }).view", "null");
shouldBe("new KeyboardEvent('eventType', { view: NaN }).view", "null");
// Note that valueOf() is not called, when the left hand side is evaluated.
shouldBeFalse("new KeyboardEvent('eventType', { view: {valueOf: function () { return window; } } }).view == window");
shouldBe("new KeyboardEvent('eventType', { get view() { return 123; } }).view", "null");
shouldThrow("new KeyboardEvent('eventType', { get view() { throw 'KeyboardEvent Error'; } })");

// keyIdentifier is passed.
// Strings.
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: 'koakuma' }).keyIdentifier", "koakuma");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: '' }).keyIdentifier", "");

// Non-strings.
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: undefined }).keyIdentifier", "undefined");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: null }).keyIdentifier", "null");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: false }).keyIdentifier", "false");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: true }).keyIdentifier", "true");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: 12345 }).keyIdentifier", "12345");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: 18446744073709551615 }).keyIdentifier", "18446744073709552000");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: NaN }).keyIdentifier", "NaN");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: [] }).keyIdentifier", "");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: [1, 2, 3] }).keyIdentifier", "1,2,3");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: {koakuma: 12345} }).keyIdentifier", "[object Object]");
shouldBeEqualToString("new KeyboardEvent('eventType', { keyIdentifier: {valueOf: function () { return 'koakuma'; } } }).keyIdentifier", "[object Object]");

// keyLocation is passed.
// numbers within the unsigned long range.
shouldBe("new KeyboardEvent('eventType', { keyLocation: 0 }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: 1 }).keyLocation", "1");
shouldBe("new KeyboardEvent('eventType', { keyLocation: 4294967294 }).keyLocation", "4294967294");
shouldBe("new KeyboardEvent('eventType', { keyLocation: 4294967295 }).keyLocation", "4294967295");

// numbers out of the unsigned long range.
// 2^{53}-1, the largest number that can be exactly represented by double.
shouldBe("new KeyboardEvent('eventType', { keyLocation: 9007199254740991 }).keyLocation", "4294967295");
// 2^{64}-1
shouldBe("new KeyboardEvent('eventType', { keyLocation: 18446744073709551615 }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: 12345678901234567890 }).keyLocation", "3944679424");
shouldBe("new KeyboardEvent('eventType', { keyLocation: -1 }).keyLocation", "4294967295");
shouldBe("new KeyboardEvent('eventType', { keyLocation: 123.45 }).keyLocation", "123");
shouldBe("new KeyboardEvent('eventType', { keyLocation: NaN }).keyLocation", "0");

// Non-numeric values.
shouldBe("new KeyboardEvent('eventType', { keyLocation: undefined }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: null }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: '' }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: '12345' }).keyLocation", "12345");
shouldBe("new KeyboardEvent('eventType', { keyLocation: '12345a' }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: 'abc' }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: [] }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: [12345] }).keyLocation", "12345");
shouldBe("new KeyboardEvent('eventType', { keyLocation: [12345, 67890] }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: {} }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: {moemoe: 12345} }).keyLocation", "0");
shouldBe("new KeyboardEvent('eventType', { keyLocation: {valueOf: function () { return 12345; }} }).keyLocation", "12345");

// ctrlKey, altKey, shiftKey and metaKey are passed.
["ctrlKey", "altKey", "shiftKey", "metaKey"].forEach(function (attr) {
    shouldBe("new KeyboardEvent('eventType', { " + attr + ": false })." + attr, "false");
    shouldBe("new KeyboardEvent('eventType', { " + attr + ": true })." + attr, "true");
});

// All initializers are passed.
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).bubbles", "true");
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).cancelable", "true");
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).view", "window");
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).detail", "111");
shouldBeEqualToString("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).keyIdentifier", "chocolate");
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).keyLocation", "222");
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).ctrlKey", "true");
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).altKey", "true");
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).shiftKey", "true");
shouldBe("new KeyboardEvent('eventType', { bubbles: true, cancelable: true, view: window, detail: 111, keyIdentifier: 'chocolate', keyLocation: 222, ctrlKey: true, altKey: true, shiftKey: true, metaKey: true }).metaKey", "true");
</script>
<script src="../../js/resources/js-test-post.js"></script>
</body>
</html>
